requirements.txt
fastapi
uvicorn
pandas
scikit-learn
tensorflow
joblib

pandas
scikit-learn
numpy
pyyaml>=6.0

dvc[all]


# DVC and DagsHub support
dvc
dvc[http]               # for DagsHub storage access
dagshub                 # for experiment tracking + integration with DagsHub

# MLflow (for local experiment tracking, optional)
mlflow

